# api-doc.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "Imsha.by REST API",
        "description": "Rest api for managing base entities in system. Rest api for managing base entities in system."
    },
    "host": "localhost:3000",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json",
        "application/javascript"
    ],
    "produces": [
        "application/json",
        "application/javascript"
    ],
    "paths": {
        "/api/mass": {
            "post": {
                "operationId": "createMassUsingPOST",
                "summary": "Create mass",
                "tags": [
                    "mass-services",
                    "Mass"
                ],
                "description": "Create mass",
                "consumes": [
                    "application/xml"
                ],
                "produces": [
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "cityId": {
                                    "type": "string"
                                },
                                "days": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                },
                                "deleted": {
                                    "type": "boolean"
                                },
                                "duration": {
                                    "type": "integer",
                                    "format": "int64"
                                },
                                "langCode": {
                                    "type": "string"
                                },
                                "notes": {
                                    "type": "string"
                                },
                                "parishId": {
                                    "type": "string"
                                },
                                "singleStartTimestamp": {
                                    "type": "integer"
                                },
                                "time": {
                                    "type": "string"
                                }
                            },
                            "required": [
                                "cityId",
                                "days",
                                "duration",
                                "langCode",
                                "parishId"
                            ],
                            "example": {
                                "cityId": "580bb8a9f9efd4680c80db8d",
                                "langCode": "RU",
                                "duration": 36004,
                                "days": [
                                    1,
                                    2,
                                    3,
                                    4
                                ],
                                "parishId": "574360fcccbd5297c86047fd",
                                "deleted": false,
                                "notes": "Test",
                                "singleStartTimestamp": 1477408490906
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Mass"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "x-stoplight": {
                    "id": "createMassUsingPOST",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "statusCode": 200,
                        "enabled": false
                    }
                }
            },
            "get": {
                "operationId": "filterMassesUsingGET",
                "summary": "Filter Masses",
                "tags": [
                    "mass-services",
                    "Mass"
                ],
                "consumes": [
                    "application/xml"
                ],
                "produces": [
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "filter",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "$ref": "#/parameters/trait:pageSort:offset"
                    },
                    {
                        "$ref": "#/parameters/trait:pageSort:sort"
                    },
                    {
                        "$ref": "#/parameters/trait:pageSort:limit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "content": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Mass"
                                    }
                                },
                                "totalPages": {
                                    "type": "integer"
                                },
                                "totalElements": {
                                    "type": "integer"
                                },
                                "last": {
                                    "type": "boolean"
                                },
                                "size": {
                                    "type": "integer"
                                },
                                "number": {
                                    "type": "integer"
                                },
                                "sort": {
                                    "type": "null"
                                },
                                "numberOfElements": {
                                    "type": "integer"
                                },
                                "first": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/responses/trait:standartErrors:400"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "$ref": "#/responses/trait:standartErrors:500"
                    }
                },
                "x-stoplight": {
                    "id": "filterMassesUsingGET",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "statusCode": 200,
                        "enabled": false
                    }
                }
            },
            "put": {
                "operationId": "PUT_api-mass",
                "summary": "Update Mass",
                "tags": [
                    "Mass"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Mass"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                },
                "x-stoplight": {
                    "id": "PUT_api-mass",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "enabled": false,
                        "dynamic": false,
                        "statusCode": 200
                    }
                }
            }
        },
        "/api/mass/{massId}": {
            "parameters": [
                {
                    "name": "massId",
                    "in": "path",
                    "description": "ID of requested mass ",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "operationId": "GET_api-mass-massId",
                "summary": "Get Mass details",
                "tags": [
                    "Mass"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Mass"
                        }
                    },
                    "400": {
                        "$ref": "#/responses/trait:standartErrors:400"
                    },
                    "403": {
                        "$ref": "#/responses/trait:standartErrors:403"
                    },
                    "404": {
                        "$ref": "#/responses/trait:standartErrors:404"
                    },
                    "500": {
                        "$ref": "#/responses/trait:standartErrors:500"
                    }
                },
                "x-stoplight": {
                    "id": "GET_api-mass-massId",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "enabled": false,
                        "dynamic": false,
                        "statusCode": 200
                    }
                }
            },
            "delete": {
                "operationId": "DELETE_api-mass-massId",
                "summary": "Delete Mass",
                "tags": [
                    "Mass"
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "$ref": "#/responses/trait:standartErrors:400"
                    },
                    "403": {
                        "$ref": "#/responses/trait:standartErrors:403"
                    },
                    "404": {
                        "$ref": "#/responses/trait:standartErrors:404"
                    },
                    "500": {
                        "$ref": "#/responses/trait:standartErrors:500"
                    }
                },
                "x-stoplight": {
                    "id": "DELETE_api-mass-massId",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "enabled": false,
                        "dynamic": false,
                        "statusCode": 200
                    }
                }
            }
        },
...
    },
    "parameters": {
        "trait:pageSort:offset": {
            "name": "offset",
            "in": "query",
            "description": "number of page",
            "type": "string",
            "default": "0"
        },
        "trait:pageSort:sort": {
            "name": "sort",
            "in": "query",
            "description": "sorting values, format: +/-[field_name]",
            "type": "string",
            "default": "+name"
        },
        "trait:pageSort:limit": {
            "name": "limit",
            "in": "query",
            "description": "size of page",
            "type": "string",
            "default": "10"
        }
    },
    "responses": {
        "trait:standartErrors:400": {
            "description": "",
            "schema": {
                "type": "object",
                "properties": {
                    "timestamp": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "error": {
                        "type": "string"
                    },
                    "exception": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "codes": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "arguments": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "codes": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            },
                                            "arguments": {
                                                "type": "null"
                                            },
                                            "defaultMessage": {
                                                "type": "string"
                                            },
                                            "code": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                },
                                "defaultMessage": {
                                    "type": "string"
                                },
                                "objectName": {
                                    "type": "string"
                                },
                                "field": {
                                    "type": "string"
                                },
                                "rejectedValue": {
                                    "type": "boolean"
                                },
                                "bindingFailure": {
                                    "type": "boolean"
                                },
                                "code": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    }
                }
            }
        },
        "trait:standartErrors:403": {
            "description": "",
            "schema": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            }
        },
        "trait:standartErrors:404": {
            "description": "",
            "schema": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    }
                }
            }
        },
        "trait:standartErrors:500": {
            "description": ""
        }
    },
    "definitions": {
        "Mass": {
            "type": "object",
            "properties": {
                "cityId": {
                    "type": "string"
                },
                "days": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "deleted": {
                    "type": "boolean"
                },
                "duration": {
                    "type": "integer",
                    "format": "int64"
                },
                "id": {
                    "type": "string"
                },
                "langCode": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "parishId": {
                    "type": "string"
                },
                "singleStartTimestamp": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                }
            },
            "required": [
                "cityId",
                "days",
                "duration",
                "langCode",
                "parishId"
            ],
            "x-stoplight": {
                "id": "Mass",
                "name": "Mass",
                "public": true
            }
        },
...
    },
...
}